Add double-buffering for concurrent flushing.
This adds a second event buffer of the same size.
On flush, the buffers are swapped and the global counter is reset.
(Both protected by the mutex.)
Flushing than exclusively works on the swapped buffer, so that tracing
can continue concurrently.
This effectively doubles the amount of memory needed for buffering.